Fix time_comparison for dates
Assume a ActiveRecord with a release date

  has_event :release, field_type: :date,
                      strategy: :time_comparison

and let's further that the release date is `Date.today`.
Previously calling `#released?` will return `false` due to [checking
with `#past?`][1]. Although, `true` might have been expected in the this
situation due to considering the release date as "the first date being
released".
The new behavior returns `true` for exactly this edge case of
date-events with time_comparison strategy by using `! #future?` instead
of '#past?`.

[1]: https://github.com/pienkowb/active_record-events/blob/2814f2d3828ab69b4cb22890a50d0a77b21ff7cd/lib/active_record/events/method_factory.rb#L42